+++ CHANGELOG.md
@@ -14,7 +14,14 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - Initial release
 
-
 ## [1.1.0] - 2023-05-01
 
 - Added support for setting pocketbase email and password in settings
+
+## [1.2.0] - 2023-05-02
+
+- Added support for setting device name in settings. If omitted, you will be prompted to provide a name each time you create a session
+
+## [1.3.0] - 2023-05-02
+
+- Added support for deleting sessions from the command palette. Note - this doesn't (currently) end the session on the device, it just removes it from the list of sessions in the DB